requirements.txt
flask-restx==1.2.0
Flask-SQLAlchemy
python-dotenv==0.21.0
deepface==0.0.79
